+++ api/V0StatusM3s.go
@@ -25,8 +25,7 @@ func V0StatusM3s(w http.ResponseWriter, r *http.Request) {
 
 	logrus.Debug("HTTP GET V0StatusM3s: ", string(d))
 
-	w.WriteHeader(http.StatusAccepted)
-	w.Header().Set("Content-Type", "text/plain; charset=utf-8")
+	w.Header().Set("Content-Type", "application/json; charset=utf-8")
 	w.Header().Set("Api-Service", "v0")
 
 	w.Write(d)