+++ api/V0UpdateBootstrap.go
@@ -39,8 +39,7 @@ func V0UpdateBootstrap(w http.ResponseWriter, r *http.Request) {
 
 	content, err := ioutil.ReadAll(res.Body)
 
-	w.WriteHeader(http.StatusAccepted)
-	w.Header().Set("Content-Type", "text/plain; charset=utf-8")
+	w.Header().Set("Content-Type", "application/json; charset=utf-8")
 	w.Header().Set("Api-Service", "v0")
 	w.Write(content)
 }