+++ types/types.go
@@ -1,40 +1,25 @@
 package types
 
-import mesosproto "github.com/AVENTER-UG/mesos-m3s/proto"
+import (
+	"context"
+
+	mesosutil "github.com/AVENTER-UG/mesos-util"
+	mesosproto "github.com/AVENTER-UG/mesos-util/proto"
+	goredis "github.com/go-redis/redis/v8"
+)
 
 // Config is a struct of the framework configuration
 type Config struct {
-	FrameworkHostname           string
-	FrameworkPort               string
-	FrameworkBind               string
-	FrameworkUser               string
-	FrameworkName               string
-	FrameworkRole               string
-	FrameworkInfo               mesosproto.FrameworkInfo
-	FrameworkInfoFile           string
-	FrameworkInfoFilePath       string
 	Principal                   string
-	Username                    string
-	Password                    string
-	MesosMasterServer           string
-	MesosSSL                    bool
-	MesosStreamID               string
-	MesosCNI                    string
-	TaskID                      uint64
-	SSL                         bool
 	LogLevel                    string
 	MinVersion                  string
 	AppName                     string
 	EnableSyslog                bool
 	Hostname                    string
 	Listen                      string
-	CommandChan                 chan Command `json:"-"`
-	State                       map[string]State
 	Domain                      string
 	K3SServerURL                string
-	K3SAgentCount               int
 	K3SAgentMax                 int
-	K3SServerCount              int
 	K3SServerMax                int
 	K3SServerPort               int
 	K3SCustomDomain             string
@@ -52,9 +37,7 @@ type Config struct {
 	VolumeDriver                string
 	VolumeK3SServer             string
 	PrefixHostname              string
-	PrefixTaskName              string
 	K3SToken                    string
-	ETCDCount                   int
 	ETCDMax                     int
 	DockerSock                  string
 	BootstrapURL                string
@@ -66,6 +49,11 @@ type Config struct {
 	ETCDMEM                     float64
 	M3SStatus                   M3SStatus
 	MesosSandboxVar             string
+	RedisServer                 string
+	RedisClient                 *goredis.Client
+	RedisCTX                    context.Context
+	RedisPassword               string
+	RedisDB                     int
 }
 
 // M3SStatus store the current TaskState of the M3s services
@@ -76,40 +64,9 @@ type M3SStatus struct {
 	Etcd   []mesosproto.TaskState
 }
 
-// Command is a chan which include all the Information about the started tasks
-type Command struct {
-	ContainerImage     string                                            `json:"container_image,omitempty"`
-	ContainerType      string                                            `json:"container_type,omitempty"`
-	TaskName           string                                            `json:"task_name,omitempty"`
-	Command            string                                            `json:"command,omitempty"`
-	Hostname           string                                            `json:"hostname,omitempty"`
-	Domain             string                                            `json:"domain,omitempty"`
-	Privileged         bool                                              `json:"privileged,omitempty"`
-	NetworkMode        string                                            `json:"network_mode,omitempty"`
-	Volumes            []mesosproto.Volume                               `protobuf:"bytes,2,rep,name=volumes" json:"volumes,omitempty"`
-	Shell              bool                                              `protobuf:"varint,6,opt,name=shell,def=1" json:"shell,omitempty"`
-	Uris               []mesosproto.CommandInfo_URI                      `protobuf:"bytes,1,rep,name=uris" json:"uris,omitempty"`
-	Environment        mesosproto.Environment                            `protobuf:"bytes,2,opt,name=environment" json:"environment,omitempty"`
-	NetworkInfo        []mesosproto.NetworkInfo                          `protobuf:"bytes,2,opt,name=networkinfo" json:"networkinfo,omitempty"`
-	DockerPortMappings []mesosproto.ContainerInfo_DockerInfo_PortMapping `protobuf:"bytes,3,rep,name=port_mappings,json=portMappings" json:"port_mappings,omitempty"`
-	DockerParameter    []mesosproto.Parameter                            `protobuf:"bytes,5,rep,name=parameters" json:"parameters,omitempty"`
-	Arguments          []string                                          `protobuf:"bytes,7,rep,name=arguments" json:"arguments,omitempty"`
-	Discovery          mesosproto.DiscoveryInfo                          `protobuf:"bytes,12,opt,name=discovery" json:"discovery,omitempty"`
-	Executor           mesosproto.ExecutorInfo
-	InternalID         int
-	TaskID             uint64
-	IsK3SAgent         bool
-	IsK3SServer        bool
-	IsETCD             bool
-	Memory             float64
-	CPU                float64
-	Agent              string
-	Labels             []mesosproto.Label
-}
-
 // State will have the state of all tasks stated by this framework
 type State struct {
-	Command Command                `json:"command"`
+	Command mesosutil.Command      `json:"command"`
 	Status  *mesosproto.TaskStatus `json:"status"`
 }
 
@@ -130,4 +87,11 @@ type Version struct {
 	BootstrapBuild string `json:"bootstrap_build"`
 	M3sBuild       string `json:"m3s_build"`
 	M3sBersion     string `json:"m3s_version"`
+}
+
+// ErrorMsg hold the structure of error messages
+type ErrorMsg struct {
+	Message  string
+	Number   int
+	Function string
 }