+++ changelog.md
@@ -2,8 +2,14 @@
 
 ## master
 
+- Add Redis Authentication Support
 - Bootstrap: Add update call to update the bootstrap server.
 - Add update api call to update the bootstrap server.
+- Fix hostname constraint method. The server/agent should never run
+  on a other worker as defined in the contrains variable.
+- Fix mesos-cli to determine framework uuid by name.
+- Move statefile to Redis DB
+
 
 ## v0.2.0